refactor(datasets): refactor timeseries graph component
